# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"address-book-controller","description":"Address Book Controller"}],"paths":{"/addressBookEntries":{"get":{"tags":["address-book-controller"],"summary":"Retrieve address book entries","description":"Retrieve all unique address book entries, regardless of address book","operationId":"getUniqueAddressBookEntriesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/AddressBookEntry"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/addressBooks":{"put":{"tags":["address-book-controller"],"summary":"Add a new address book","operationId":"addAddressBookUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"addressBook","description":"addressBook","required":true,"schema":{"$ref":"#/definitions/AddressBook"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AddressBook"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/addressBooks/{addressBookId}":{"get":{"tags":["address-book-controller"],"summary":"Retrieve address book","description":"Retrieve a single address book for a given id, including all entries","operationId":"getAddressBookUsingGET","produces":["*/*"],"parameters":[{"name":"addressBookId","in":"path","description":"addressBookId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AddressBook"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/addressBooks/{addressBookId}/entry":{"put":{"tags":["address-book-controller"],"summary":"Add an address book entry","operationId":"addAddressBookEntryUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"addressBookEntry","description":"addressBookEntry","required":true,"schema":{"$ref":"#/definitions/AddressBookEntry"}},{"name":"addressBookId","in":"path","description":"addressBookId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AddressBookEntry"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/addressBooks/{addressBookId}/entry/{addressBookEntryId}":{"delete":{"tags":["address-book-controller"],"summary":"Delete an address book entry","operationId":"deleteAddressBookEntryUsingDELETE","produces":["*/*"],"parameters":[{"name":"addressBookEntryId","in":"path","description":"addressBookEntryId","required":true,"type":"integer","format":"int64"},{"name":"addressBookId","in":"path","description":"addressBookId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"AddressBook":{"type":"object","properties":{"addressBookEntries":{"type":"array","items":{"$ref":"#/definitions/AddressBookEntry"}},"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"AddressBook"},"AddressBookEntry":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"phoneNumber":{"type":"string"}},"title":"AddressBookEntry"}}}